% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_methods_tbl.R
\name{dist_methods_tbl_cat}
\alias{dist_methods_tbl_cat}
\title{List available categorical dissimilarities}
\usage{
dist_methods_tbl_cat()
}
\value{
A tibble containing the rows of [dist_methods_tbl()] with
  `argument == "method_cat"`.
}
\description{
Convenience wrapper around [dist_methods_tbl()] returning only methods that
can be used as categorical dissimilarities through `method_cat`.
}
\examples{
dist_methods_tbl_cat()

}
