% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{generate toy mixed datasets for the supplementary material: figures 1 to 3}
\usage{
generate_dataset(
  n,
  porig,
  pn,
  pnnoise,
  pcnoise,
  sigma,
  qoptions,
  seed = NULL,
  mode = "per_variable"
)
}
\arguments{
\item{n}{number of observations}

\item{porig}{number of original informative continuous variables}

\item{pn}{number of continuous variables (total, before adding noise variables)}

\item{pnnoise}{number of extra numeric noise variables}

\item{pcnoise}{number of extra categorical noise variables}

\item{sigma}{sd for noise added to informative variables}

\item{qoptions}{number of bins for categorization (vector if per_variable, scalar if shared)}

\item{seed}{optional seed}

\item{mode}{either "per_variable" or "shared"}
}
\description{
generate toy mixed datasets for the supplementary material: figures 1 to 3
}
