% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdist_summary_impl.R
\name{mdist_summary_impl}
\alias{mdist_summary_impl}
\title{Internal: summary method implementation for MDist R6 objects}
\usage{
mdist_summary_impl(object, ...)
}
\arguments{
\item{object}{An `MDist` object to summarize.}

\item{...}{Additional arguments currently not used.}
}
\value{
Invisibly returns `object`.
}
\description{
Internal: summary method implementation for MDist R6 objects
}
