% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam_dist_functions.R
\name{pam_dist}
\alias{pam_dist}
\title{PAM clustering specification based on manydist dissimilarities}
\usage{
pam_dist(num_clusters = NULL)
}
\arguments{
\item{num_clusters}{Number of clusters.}
}
\value{
A `pam_dist_spec` object.
}
\description{
PAM clustering specification based on manydist dissimilarities
}
