% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_dist_functions.R
\name{predict_knn_dist_class}
\alias{predict_knn_dist_class}
\title{Predict from a KNN-distance model (classification)}
\usage{
predict_knn_dist_class(object, new_data, type = c("class", "prob"))
}
\arguments{
\item{object}{A `knn_dist` object created by `fit_knn_dist()`.}

\item{new_data}{A data frame or matrix of precomputed distances
*or* raw predictors, depending on `dist_fun`.}

\item{type}{`"class"` or `"prob"`.}
}
\description{
Used by parsnip for `type = "class"` and `type = "prob"`.
}
