% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_dist_functions.R
\name{predict_knn_dist_prob}
\alias{predict_knn_dist_prob}
\title{Predict class probabilities from a KNN-distance model}
\usage{
predict_knn_dist_prob(object, new_data, ...)
}
\arguments{
\item{object}{A `knn_dist` object created by `fit_knn_dist()`.}

\item{new_data}{A data frame or matrix of precomputed distances
*or* raw predictors, depending on `dist_fun`.}
}
\description{
Thin wrapper used by parsnip for `type = "prob"`.
}
