% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_methods_tbl.R
\name{response_aware_methods}
\alias{response_aware_methods}
\title{List response-aware methods}
\usage{
response_aware_methods(data_type = NULL, argument = NULL)
}
\arguments{
\item{data_type}{Optional character vector used to restrict the returned
methods by data type, for example `"categorical"` or `"mixed"`.}

\item{argument}{Optional character vector used to restrict the returned
methods by argument, for example `"method_cat"` or `"preset"`.}
}
\value{
A character vector of response-aware method names.
}
\description{
Returns the methods in [dist_methods_tbl()] that can use a response variable.
The output can optionally be restricted by data type or by `mdist()`
argument.
}
\examples{
response_aware_methods()
response_aware_methods(argument = "method_cat")
response_aware_methods(argument = "preset")

}
