% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_dist_functions.R
\name{spectral_dist}
\alias{spectral_dist}
\title{Spectral clustering specification based on manydist dissimilarities}
\usage{
spectral_dist(num_clusters = NULL, sigma = NULL, nstart = 50)
}
\arguments{
\item{num_clusters}{Number of clusters.}

\item{sigma}{Optional bandwidth for the Gaussian affinity. If `NULL`,
the median pairwise distance is used.}

\item{nstart}{Number of random starts for k-means.}
}
\value{
A `spectral_dist_spec` object.
}
\description{
Spectral clustering specification based on manydist dissimilarities
}
