% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_from_dist.R
\name{spectral_from_dist}
\alias{spectral_from_dist}
\title{Spectral clustering from a distance matrix}
\usage{
spectral_from_dist(D, k, affinity_method = "selftune")
}
\arguments{
\item{D}{A distance matrix}

\item{k}{Number of clusters}

\item{affinity_method}{Method to build affinity}
}
\value{
A vector of cluster labels
}
\description{
Spectral clustering from a distance matrix
}
