% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_dist_tunable.R
\name{tunable.nearest_neighbor_dist}
\alias{tunable.nearest_neighbor_dist}
\title{Tuning parameters for nearest_neighbor_dist}
\usage{
\method{tunable}{nearest_neighbor_dist}(x, ...)
}
\arguments{
\item{x}{A `nearest_neighbor_dist` model specification.}

\item{...}{Not used (included for method consistency).}
}
\description{
This defines which arguments of `nearest_neighbor_dist()` are tunable
and which dials parameter to use for each one.
}
