% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdist.R
\name{mdist}
\alias{mdist}
\title{Mixed-type distance for mixed data}
\usage{
mdist(
  x,
  new_data = NULL,
  response = NULL,
  distance_cont = "manhattan",
  distance_cat = "tot_var_dist",
  commensurable = FALSE,
  scaling_cont = "none",
  ncomp = NULL,
  threshold = NULL,
  preset = "custom",
  interaction = FALSE,
  prop_nn = 0.1,
  score = "ba",
  decision = "prior_corrected"
)
}
\arguments{
\item{x}{A data frame or matrix of predictors.}

\item{new_data}{Optional new data to compute distances to \code{x}.}

\item{response}{Optional response variable used for supervised
categorical distances.}

\item{distance_cont}{Character string specifying the distance
for continuous variables.}

\item{distance_cat}{Character string specifying the distance
for categorical variables.}

\item{commensurable}{Logical; whether to apply commensurability
scaling between variable types.}

\item{scaling_cont}{Character string specifying scaling for
continuous variables.}

\item{ncomp}{Optional number of components used in dimensional
reduction. If \code{NULL}, no reduction is applied.}

\item{threshold}{Optional threshold parameter.}

\item{preset}{Character string specifying a predefined setup.}

\item{interaction}{Logical; whether to include interaction-based
distances.}

\item{prop_nn}{proportion of neighbours to consider when measuring interactions}

\item{score}{classification metric either  "ba" (balanced accuracy) or "logloss"}

\item{decision}{rule when score is set to ba}
}
\value{
A dissimilarity object.
}
\description{
Computes a distance or dissimilarity object for mixed-type data,
combining continuous and categorical variables.
}
