% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_dist_parsnip_registration.R
\name{nearest_neighbor_dist}
\alias{nearest_neighbor_dist}
\title{Nearest neighbor with precomputed distances}
\usage{
nearest_neighbor_dist(mode = "classification", neighbors = NULL)
}
\arguments{
\item{mode}{"classification" or "regression".}

\item{neighbors}{Number of neighbors (k).}
}
\description{
Nearest neighbor with precomputed distances
}
