% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mdist_functions.R
\name{step_mdist}
\alias{step_mdist}
\title{Distance-based representation via manydist}
\usage{
step_mdist(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  preset = "custom",
  distance_cont = "manhattan",
  distance_cat = "tot_var_dist",
  commensurable = FALSE,
  scaling_cont = "none",
  ncomp = NULL,
  threshold = NULL,
  columns = NULL,
  train_predictors = NULL,
  skip = FALSE,
  id = recipes::rand_id("mdist")
)
}
\arguments{
\item{recipe}{A recipe object.}

\item{...}{Selector(s) for the predictor columns to be used in mdist.}

\item{role}{Role for the new distance columns (usually "predictor").}

\item{trained}{Logical for recipes internals (do not set manually).}

\item{preset, distance_cont, distance_cat, commensurable, scaling_cont, }{ncomp,threshold Arguments forwarded to `mdist()`.}

\item{columns}{Names of columns selected at prep time (recipes internals).}

\item{train_predictors}{Training predictors stored at prep time (internals).}

\item{skip, id}{Standard recipes arguments.}
}
\description{
Distance-based representation via manydist
}
